Move linux/GPUMeter.* to GPUMeter.*
The GPUMeter code is now platform independent.
