+++ Makefile.am
@@ -119,6 +119,7 @@ myhtopheaders = \
 	EnvScreen.h \
 	FileDescriptorMeter.h \
 	FunctionBar.h \
+	GPUMeter.h \
 	Hashtable.h \
 	Header.h \
 	HeaderLayout.h \
@@ -174,7 +175,6 @@ linux_platform_headers = \
 	generic/uname.h \
 	linux/CGroupUtils.h \
 	linux/GPU.h \
-	linux/GPUMeter.h \
 	linux/HugePageMeter.h \
 	linux/IOPriority.h \
 	linux/IOPriorityPanel.h \
@@ -195,12 +195,12 @@ linux_platform_headers = \
 	zfs/ZfsCompressedArcMeter.h
 
 linux_platform_sources = \
+	GPUMeter.c \
 	generic/gettime.c \
 	generic/hostname.c \
 	generic/uname.c \
 	linux/CGroupUtils.c \
 	linux/GPU.c \
-	linux/GPUMeter.c \
 	linux/HugePageMeter.c \
 	linux/IOPriorityPanel.c \
 	linux/LibSensors.c \
